ALSA: hda: intel-nhlt: add intel_nhlt_ssp_dir_mask() function
This helper function returns bitmask of SSP ports which support
specific direction according to endpoint descriptors. Smart
amplifier's render feedback does not count when finding SSP ports with
capture direction so caller could use this function to find out which
port is for headset(both render and capture) and which port is for
speakers(render only).

Signed-off-by: Brent Lu <brent.lu@intel.com>